automation/transcriber/requirements.txt
SpeechRecognition==3.10.4
moviepy==1.0.3
pydub==0.25.1
openai-whisper==20240930